+++ packages/components/src/box-control/types.ts
@@ -1,9 +1,4 @@
 /**
- * External dependencies
- */
-import type { useHover } from '@use-gesture/react';
-
-/**
  * Internal dependencies
  */
 import type { UnitControlProps } from '../unit-control/types';
@@ -14,6 +9,10 @@ export type BoxControlValue = {
 	right?: string;
 	bottom?: string;
 	left?: string;
+};
+
+export type CustomValueUnits = {
+	[ key: string ]: { max: number; step: number };
 };
 
 type UnitControlPassthroughProps = Omit<
@@ -91,22 +90,6 @@ export type BoxControlInputControlProps = UnitControlPassthroughProps & {
 	sides: BoxControlProps[ 'sides' ];
 	values: BoxControlValue;
 };
-
-export type BoxUnitControlProps = UnitControlPassthroughProps &
-	Pick< UnitControlProps, 'onChange' | 'onFocus' > & {
-		isFirst?: boolean;
-		isLast?: boolean;
-		isOnly?: boolean;
-		label?: string;
-		onHoverOff?: (
-			event: ReturnType< typeof useHover >[ 'event' ],
-			state: Omit< ReturnType< typeof useHover >, 'event' >
-		) => void;
-		onHoverOn?: (
-			event: ReturnType< typeof useHover >[ 'event' ],
-			state: Omit< ReturnType< typeof useHover >, 'event' >
-		) => void;
-	};
 
 export type BoxControlIconProps = {
 	/**