+++ packages/dataviews/src/add-filter.js
@@ -23,6 +23,7 @@ const {
 	DropdownMenuRadioItemV2: DropdownMenuRadioItem,
 	DropdownMenuSeparatorV2: DropdownMenuSeparator,
 	DropdownMenuItemLabelV2: DropdownMenuItemLabel,
+	DropdownMenuItemHelpTextV2: DropdownMenuItemHelpText,
 } = unlock( componentsPrivateApis );
 
 function WithSeparators( { children } ) {
@@ -120,7 +121,7 @@ export default function AddFilter( { filters, view, onChangeView } ) {
 													name={ `add-filter-${ filter.field }` }
 													value={ element.value }
 													checked={ isActive }
-													onClick={ () => {
+													onChange={ ( e ) => {
 														onChangeView( {
 															...view,
 															page: 1,
@@ -132,7 +133,9 @@ export default function AddFilter( { filters, view, onChangeView } ) {
 																		activeOperator,
 																	value: isActive
 																		? undefined
-																		: element.value,
+																		: e
+																				.target
+																				.value,
 																},
 															],
 														} );
@@ -141,6 +144,13 @@ export default function AddFilter( { filters, view, onChangeView } ) {
 													<DropdownMenuItemLabel>
 														{ element.label }
 													</DropdownMenuItemLabel>
+													{ !! element.description && (
+														<DropdownMenuItemHelpText>
+															{
+																element.description
+															}
+														</DropdownMenuItemHelpText>
+													) }
 												</DropdownMenuRadioItemCustom>
 											);
 										} ) }