+++ packages/edit-site/src/components/page-patterns/use-patterns.js
@@ -184,29 +184,38 @@ const selectPatterns = createSelector(
 	]
 );
 
-const patternBlockToPattern = ( patternBlock, categories ) => ( {
-	blocks: parse( patternBlock.content.raw, {
+/**
+ * Converts a post of type `wp_block` to a 'pattern item' that more closely
+ * matches the structure of theme provided patterns.
+ *
+ * @param {Object} patternPost The `wp_block` record being normalized.
+ * @param {Map}    categories  A Map of user created categories.
+ *
+ * @return {Object} The normalized item.
+ */
+const convertPatternPostToItem = ( patternPost, categories ) => ( {
+	blocks: parse( patternPost.content.raw, {
 		__unstableSkipMigrationLogs: true,
 	} ),
-	...( patternBlock.wp_pattern_category.length > 0 && {
-		categories: patternBlock.wp_pattern_category.map(
+	...( patternPost.wp_pattern_category.length > 0 && {
+		categories: patternPost.wp_pattern_category.map(
 			( patternCategoryId ) =>
 				categories && categories.get( patternCategoryId )
 					? categories.get( patternCategoryId ).slug
 					: patternCategoryId
 		),
 	} ),
-	termLabels: patternBlock.wp_pattern_category.map( ( patternCategoryId ) =>
+	termLabels: patternPost.wp_pattern_category.map( ( patternCategoryId ) =>
 		categories?.get( patternCategoryId )
 			? categories.get( patternCategoryId ).label
 			: patternCategoryId
 	),
-	id: patternBlock.id,
-	name: patternBlock.slug,
-	syncStatus: patternBlock.wp_pattern_sync_status || PATTERN_SYNC_TYPES.full,
-	title: patternBlock.title.raw,
-	type: PATTERN_TYPES.user,
-	patternBlock,
+	id: patternPost.id,
+	name: patternPost.slug,
+	syncStatus: patternPost.wp_pattern_sync_status || PATTERN_SYNC_TYPES.full,
+	title: patternPost.title.raw,
+	type: patternPost.type,
+	patternPost,
 } );
 
 const selectUserPatterns = createSelector(
@@ -215,7 +224,7 @@ const selectUserPatterns = createSelector(
 			select( coreStore );
 
 		const query = { per_page: -1 };
-		const records = getEntityRecords(
+		const patternPosts = getEntityRecords(
 			'postType',
 			PATTERN_TYPES.user,
 			query
@@ -225,9 +234,9 @@ const selectUserPatterns = createSelector(
 		userPatternCategories.forEach( ( userCategory ) =>
 			categories.set( userCategory.id, userCategory )
 		);
-		let patterns = records
-			? records.map( ( record ) =>
-					patternBlockToPattern( record, categories )
+		let patterns = patternPosts
+			? patternPosts.map( ( record ) =>
+					convertPatternPostToItem( record, categories )
 			  )
 			: EMPTY_PATTERN_LIST;