+++ packages/edit-site/src/components/sidebar-edit-mode/plugin-sidebar/index.js
@@ -2,8 +2,6 @@
  * WordPress dependencies
  */
 import { ComplementaryArea } from '@wordpress/interface';
-import { useSelect } from '@wordpress/data';
-import { store as preferencesStore } from '@wordpress/preferences';
 
 /**
  * Renders a sidebar when activated. The contents within the `PluginSidebar` will appear as content within the sidebar.
@@ -71,21 +69,11 @@ import { store as preferencesStore } from '@wordpress/preferences';
  * ```
  */
 export default function PluginSidebarEditSite( { className, ...props } ) {
-	const showIconLabels = useSelect(
-		( select ) =>
-			!! select( preferencesStore ).get(
-				'core/edit-site',
-				'showIconLabels'
-			),
-		[]
-	);
-
 	return (
 		<ComplementaryArea
 			panelClassName={ className }
 			className="edit-site-sidebar-edit-mode"
 			scope="core/edit-site"
-			showIconLabels={ showIconLabels }
 			{ ...props }
 		/>
 	);