+++ packages/patterns/src/private-apis.js
@@ -2,8 +2,14 @@
  * Internal dependencies
  */
 import { lock } from './lock-unlock';
-import CreatePatternModal from './components/create-pattern-modal';
-import DuplicatePatternModal from './components/duplicate-pattern-modal';
+import {
+	default as CreatePatternModal,
+	CreatePatternModalContents,
+} from './components/create-pattern-modal';
+import {
+	default as DuplicatePatternModal,
+	useDuplicatePatternProps,
+} from './components/duplicate-pattern-modal';
 import RenamePatternModal from './components/rename-pattern-modal';
 import PatternsMenuItems from './components';
 import RenamePatternCategoryModal from './components/rename-pattern-category-modal';
@@ -20,7 +26,9 @@ import {
 export const privateApis = {};
 lock( privateApis, {
 	CreatePatternModal,
+	CreatePatternModalContents,
 	DuplicatePatternModal,
+	useDuplicatePatternProps,
 	RenamePatternModal,
 	PatternsMenuItems,
 	RenamePatternCategoryModal,